chore: rename package-graph files
