+++ packages/migration-graph/tests/package-dependency-graph.test.ts
@@ -15,8 +15,8 @@ import rimraf from 'rimraf';
 import { dirSync, setGracefulCleanup } from 'tmp';
 import merge from 'lodash.merge';
 import { ProjectGraph } from '../src/project-graph';
-import { EmberAppPackageGraph } from '../src/package-graph/ember-app';
-import { EmberAddonPackageGraph } from '../src/package-graph/ember-addon';
+import { EmberAppPackageGraph } from '../src/package-graph/ember-app-package-graph';
+import { EmberAddonPackageGraph } from '../src/package-graph/ember-addon-package-graph';
 import { createPackageDependencyGraph } from '../src/package-graph';
 import { ModuleNode, PackageNode } from '../src/types';
 import { Graph } from '../src/utils/graph';
